src/data_fetch_from_api/alpha_vantage_api ( finance stocks)/requirements.txt
alpha_vantage
google-cloud-storage
google-auth
gcsfs
fsspec
pyspark
pandas
google-cloud
datetime
